add: вынес создание точек в файл point_selection_tool change: Отредактировал river.py с учётом изменнений
